fix:  reusability #68
