+++ scripts/deploy/setup-salesforce-scratch-org.js
@@ -1,6 +1,7 @@
 const sh = require('shelljs');
 const chalk = require('chalk');
 const log = console.log;
+const { assignPermissionset } = require('./util');
 
 /*
  * Create a scratch org and save user login details
@@ -50,19 +51,7 @@ const setupScratchOrg = async () => {
     }
 
     // Assign permission set to user
-    const assignPermissionset = JSON.parse(
-        sh.exec(
-            `sfdx force:user:permset:assign --permsetname Salesforce_Slack_App_Admin -u ${sh.env.SF_USERNAME} --json`,
-            { silent: true }
-        )
-    );
-
-    if (!assignPermissionset.result.successes) {
-        console.error(
-            'Permission set assignment failed - try again later: ' +
-                JSON.stringify(assignPermissionset)
-        );
-    }
+    await assignPermissionset();
 
     log(chalk.green('*** ✔ Done with the Salesforce scratch org setup'));
 };
